Move and options to the option holder module
