+++ src/aiida_vasp/workchains/v2/relax.py
@@ -25,7 +25,6 @@ CHANGELOG
 """
 
 # pylint: disable=attribute-defined-outside-init
-from typing import Optional
 
 import numpy as np
 from aiida import orm
@@ -35,10 +34,9 @@ from aiida.common.utils import classproperty
 from aiida.engine import ToContext, WorkChain, append_, if_, while_
 from aiida.orm.nodes.data.base import to_aiida_type
 from aiida.plugins import WorkflowFactory
-from pydantic import Field
 
 from aiida_vasp.utils.aiida_utils import get_data_class
-from aiida_vasp.utils.opthold import OptionContainer
+from aiida_vasp.utils.opthold import RelaxOptions
 from aiida_vasp.utils.workchains import compose_exit_code
 
 from .common import OVERRIDE_NAMESPACE, nested_update_dict_node, site_magnetization_to_magmom
@@ -49,96 +47,6 @@ __version__ = '0.5.0'
 # Change log
 # 0.4.0 update such `vasp` namespace in `parameters` is renamed to `incar`
 # 0.5.0 update the logic of convergence checking. Cell comparsion is always done using the input/output structures.
-
-
-class RelaxOptions(OptionContainer):
-    """Options for VaspRelaxWorkChain"""
-
-    algo: str = Field(description='The algorithm to use for relaxation', examples=['cg', 'rd'], default='cg')
-    energy_cutoff: Optional[float] = Field(
-        description='The cut off energy difference when the relaxation is stopped (e.g. EDIFF)',
-        default=None,
-    )
-    force_cutoff: float = Field(
-        description='The maximum force when the relaxation is stopped (e.g. EDIFFG)',
-        default=0.03,
-    )
-    steps: float = Field(description='Number of relaxation steps to perform (eg. NSW)', default=60)
-    positions: bool = Field(description='If True, perform relaxation of the atomic positions', default=True)
-    shape: bool = Field(description='If True, perform relaxation of the cell shape', default=True)
-    volume: bool = Field(description='If True, perform relaxation of the cell volume', default=True)
-    convergence_on: bool = Field(description='If True, perform convergence checks within the workchain', default=True)
-    convergence_absolute: bool = Field(
-        description='If True, use absolute values where possible when performing convergence checks',
-        default=False,
-    )
-    convergence_max_iterations: int = Field(description='Maximum iterations for convergence checking', default=5)
-    convergence_positions: float = Field(
-        description=(
-            'The cutoff value for the convergence check on positions in Angstram.'
-            ' A negative value by pass the check.'
-        ),
-        default=0.1,
-    )
-    convergence_volume: float = Field(
-        description='The cutoff value for the convergence check on volume between the two structures.'
-        ' A negative value by pass the check.',
-        default=0.01,
-    )
-    convergence_shape_lengths: float = Field(
-        description='The cutoff value for the convergence check on the lengths of the unit cell'
-        ' vectors, between input and the outputs structure. A negative value by pass'
-        ' the check.',
-        default=0.1,
-    )
-    convergence_shape_angles: float = Field(
-        description='The cutoff value for the convergence check on the angles of the unit cell vectors, '
-        'between input and the outputs structure. A'
-        ' negative value by pass the check.',
-        default=0.1,
-    )
-    convergence_mode: str = Field(
-        description="Mode of the convergence check for positions. 'inout' for checking input/output structure, "
-        "or 'last' to check only the change of"
-        ' the last step.',
-        examples=['inout', 'last'],
-        default='last',
-    )
-    reuse: bool = Field(
-        description='Whether reuse the previous calculation by copying over the remote folder',
-        default=False,
-    )
-    clean_reuse: bool = Field(
-        description='Whether to perform a final cleaning of the reused calculations', default=True
-    )
-    keep_sp_workdir: bool = Field(
-        description='Whether to keep the workdir of the final singlepoint calculation', default=False
-    )
-    perform: bool = Field(description="Do not perform any relaxation if set to 'False'", default=True)
-    hybrid_calc_bootstrap: bool = Field(
-        description='Whether to bootstrap hybrid calculation by performing standard DFT first', default=False
-    )
-    hybrid_calc_bootstrap_wallclock: int = Field(
-        description='Wall time limit in second for the bootstrap calculation', default=3600
-    )
-    keep_magnetization: bool = Field(
-        description='Whether to keep magnetization from the previous calculation if possible', default=False
-    )
-
-    # TODO: implement pyandtic checks
-
-    # @classmethod
-    # def validate_dict(cls, input_dict, port=None):
-    #     """Check mutually exclusive fields"""
-    #     super().validate_dict(input_dict, port)
-    #     if isinstance(input_dict, orm.Dict):
-    #         input_dict = input_dict.get_dict()
-    #     force_cut = input_dict.get('force_cutoff')
-    #     energy_cut = input_dict.get('energy_cutoff')
-    #     if force_cut is None and energy_cut is None:
-    #         raise InputValidationError("Either 'force_cutoff' or 'energy_cutoff' should be supplied")
-    #     if (force_cut is not None) and (energy_cut is not None):
-    #         raise InputValidationError("Cannot set both 'force_cutoff' and 'energy_cutoff'")
 
 
 class VaspRelaxWorkChain(WorkChain, WithVaspInputSet):
